+++ packages/serinus/lib/serinus.dart
@@ -1,7 +1,6 @@
 library serinus;
 
 export 'src/adapters/adapters.dart';
-export 'src/body_size_limit.dart';
 export 'src/containers/module_container.dart';
 export 'src/contexts/contexts.dart';
 export 'src/core/core.dart';
@@ -9,6 +8,7 @@ export 'src/enums/enums.dart';
 export 'src/errors/initialization_error.dart';
 export 'src/exceptions/exceptions.dart';
 export 'src/global_prefix.dart';
+export 'src/hooks/body_size_limit_hook.dart';
 export 'src/http/http.dart';
 export 'src/mixins/mixins.dart';
 export 'src/versioning.dart';