+++ packages/serinus/lib/src/core/route.dart
@@ -1,5 +1,6 @@
 import '../contexts/request_context.dart';
 import '../enums/http_method.dart';
+import '../http/http.dart';
 
 /// The [Route] class is used to define the routes of the application.
 class Route {
@@ -27,11 +28,6 @@ class Route {
   /// It can be overridden if needed.
   Future<void> transform(RequestContext context) async {}
 
-  /// The [parse] method is used to parse and validate the request context.
-  ///
-  /// It can be overridden if needed.
-  Future<void> parse(RequestContext context) async {}
-
   /// The [beforeHandle] method is used to execute code before the route is handled.
   ///
   /// It can be overridden if needed.
@@ -40,7 +36,7 @@ class Route {
   /// The [afterHandle] method is used to execute code after the route is handled.
   ///
   /// It can be overridden if needed.
-  Future<void> afterHandle(RequestContext context) async {}
+  Future<void> afterHandle(RequestContext context, Response response) async {}
 
   /// The [Route.get] factory constructor is used to create a new instance of the [Route] class with the GET method.
   factory Route.get(String path) {