+++ packages/serinus/lib/src/http/http.dart
@@ -1,5 +1,4 @@
 export 'body.dart';
-export 'cors.dart';
 export 'form_data.dart';
 export 'internal_response.dart';
 export 'request.dart';