+++ packages/serinus/lib/src/http/request.dart
@@ -1,12 +1,9 @@
-import 'dart:convert';
 import 'dart:io';
 
 import '../extensions/content_type_extensions.dart';
 import '../extensions/string_extensions.dart';
-import 'body.dart';
-import 'form_data.dart';
+import 'http.dart';
 import 'internal_request.dart';
-import 'session.dart';
 
 /// The class [Request] is used to create a request object.
 ///
@@ -51,6 +48,9 @@ class Request {
   /// The path of the request.
   String get path => _original.path;
 
+  /// The uri of the request.
+  Uri get uri => _original.uri;
+
   /// The method of the request.
   String get method => _original.method;
 
@@ -58,7 +58,7 @@ class Request {
   Map<String, dynamic> get headers => _original.headers;
 
   /// The query parameters of the request.
-  Map<String, dynamic> get queryParameters => _queryParamters;
+  Map<String, dynamic> get query => _queryParamters;
 
   /// The session of the request.
   Session get session => Session(_original.original.session);
@@ -74,8 +74,21 @@ class Request {
 
   final Map<String, dynamic> _data = {};
 
+  /// The operator [] is used to get data from the request.
+  dynamic operator [](String key) => _data[key];
+
+  /// The operator []= is used to set data to the request.
+  void operator []=(String key, dynamic value) {
+    _data[key] = value;
+  }
+
   /// The body of the request.
   Body? body;
+
+  /// The content type of the request.
+  int get contentLength => _original.contentLength > -1
+      ? _original.contentLength
+      : body?.length ?? 0;
 
   /// This method is used to parse the body of the request.
   ///
@@ -111,8 +124,9 @@ class Request {
 
     /// If the content type is json, it will parse the body as a json object.
     final parsedJson = parsedBody.tryParse();
-    if (parsedJson != null || contentType == ContentType.json) {
-      final json = parsedJson ?? jsonDecode(parsedBody);
+    if ((parsedJson != null && contentType == ContentType.json) ||
+        parsedJson != null) {
+      final json = parsedJson;
       body = Body(contentType, json: json);
       return;
     }