+++ packages/serinus_cli/lib/src/commands/commands.dart
@@ -1,1 +1,2 @@
+export 'deploy_command.dart';
 export 'update_command.dart';